+++ backend/src/Controller.ts
@@ -52,6 +52,7 @@ class Controller {
     if (!socket) Logger.error('Socket.io server failed to start.')
     if (!socket) throw new Error('Socket.io server failed to start.')
     Logger.info('OPEN SOCKETS', { existing: socket.eventNames() })
+    this.checkBackendSetting()
     if (socket.eventNames().includes('init')) socket.removeAllListeners()
 
     socket.on('user/lock', user.signOut)
@@ -80,7 +81,6 @@ class Controller {
     socket.on('heartbeat', this.check)
     socket.on('showFolder', this.showFolder)
     socket.on('maximize', () => EventBus.emit(electronInterface.EVENTS.maximize))
-    socket.on('backend/check-setting', this.checkBackendSetting)
 
     this.initBackend()
     this.check()
