+++ frontend/src/services/graphQL.ts
@@ -15,9 +15,9 @@ export async function graphQLBasicRequest(query: String, variables: ILookup<any>
 
 export async function graphQLRequest(query: String, variables: ILookup<any> = {}) {
   if (store.getState().ui.offline) return {}
-  const state = store.getState()
+  const { backend } = store.getState()
 
-  const { apiURL, betaApiURL } = state.backend.backendSetting
+  const { apiURL, betaApiURL } = backend.environment.backendSetting
 
   const token = await getToken()
